+++ docker-entrypoint.sh
@@ -9,4 +9,8 @@ if [ "$1" = "migrate" ]; then
   exec node /app/server/cli/migrate-to-single-images.mjs "${@:2}"
 fi
 
+if [ "$1" = "slider" ]; then
+  exec node /app/server/cli/slider.mjs "${@:2}"
+fi
+
 exec node /app/.output/server/index.mjs "$@"